+++ example/index.tsx
@@ -6,6 +6,12 @@ import Modal from '../.';
 const App = () => {
   const [visible, setVisible] = React.useState(false);
   const [secondVisible, setSecondVisible] = React.useState(false);
+
+  const handleClick = () => {
+    console.log(111)
+    setVisible(true)
+  }
+
   const handleSecond = () => {
     setSecondVisible(true)
   }
@@ -19,7 +25,7 @@ const App = () => {
 
   return (
     <div style={{ height: 5000 }}>
-      <button style={{ width: 50, height: 30 }} onClick={() => setVisible(true)}>Open</button>
+      <button style={{ width: 50, height: 30 }} onClick={handleClick}>Open</button>
       <Modal
         title={<div>Test Title</div>}
         visible={visible}
@@ -30,6 +36,10 @@ const App = () => {
         maskClosable={true}
         escClosable={false}
         canFullscreen={false}
+        centered={true}
+        footer={null}
+        zIndex={1050}
+        animated={true}
       >
         <button onClick={handleSecond}>more modal</button>
       </Modal>