+++ src/index.tsx
@@ -4,13 +4,16 @@ import Modal from './components/modal';
 
 export interface SModalProps {
   afterClose?: () => void,
+  animated?: boolean,
   borderRadius?: number | string,
   cancelText?: string,
   canFullscreen?: boolean,
+  centered?: boolean,
   children?: React.ReactNode,
   closable?: boolean,
   draggable?: boolean,
   escClosable?: boolean,
+  footer?: React.ReactNode,
   maskClosable?: boolean
   onCancel?: () => void,
   onSure?: () => void,
@@ -18,18 +21,22 @@ export interface SModalProps {
   title?: React.ReactNode | string,
   visible?: boolean,
   width?: number | string,
+  zIndex?: number
 }
 
 const LModal: React.FC<SModalProps> = (props) => {
   const {
     afterClose,
+    animated = true,
     borderRadius = 5,
     cancelText = '取消',
     canFullscreen = false,
+    centered = false,
     children = <></>,
     closable = true,
     draggable = false,
     escClosable = true,
+    footer,
     maskClosable = true,
     onCancel,
     onSure,
@@ -37,6 +44,7 @@ const LModal: React.FC<SModalProps> = (props) => {
     title = 'Modal',
     visible = false,
     width = 500,
+    zIndex = 1001
   } = props;
   const body: any = document.querySelector('body');
   // 是否初始化过后已经被调用过了
@@ -53,20 +61,29 @@ const LModal: React.FC<SModalProps> = (props) => {
       // 判断是否初始化过后已经被调用过了
       if (!isInit) setIsInit(true)
       setIsShow(true)
-      // body.style.overflow = 'hidden'
-      setTransitionClass('modal-fade-enter modal-fade-enter-active')
-      if (enterTimeout) clearTimeout(enterTimeout)
-      setEnterTimeout(setTimeout(() => {
-        setTransitionClass('')
-      }, 200))
+      if (animated) {
+        setTransitionClass('modal-fade-enter modal-fade-enter-active')
+        if (enterTimeout) clearTimeout(enterTimeout)
+        setEnterTimeout(setTimeout(() => {
+          setTransitionClass('')
+        }, 200))
+      }
     } else {
-      setTransitionClass('modal-fade-leave modal-fade-leave-active')
-      if (leaveTimeout) clearTimeout(leaveTimeout)
-      setLeaveTimeout(setTimeout(() => {
-        setTransitionClass('')
-        setIsShow(false)
-        if (isInit && afterClose) afterClose()
-      }, 200))
+      if (isInit) {
+        if (animated) {
+          setTransitionClass('modal-fade-leave modal-fade-leave-active')
+          if (enterTimeout) clearTimeout(enterTimeout)
+          if (leaveTimeout) clearTimeout(leaveTimeout)
+          setLeaveTimeout(setTimeout(() => {
+            setTransitionClass('')
+            setIsShow(false)
+            if (isInit && afterClose) afterClose()
+          }, 200))
+        } else {
+          setIsShow(false)
+          if (isInit && afterClose) afterClose()
+        }
+      }
     }
     return () => {
       // 清除timeout
@@ -82,10 +99,12 @@ const LModal: React.FC<SModalProps> = (props) => {
       borderRadius={borderRadius}
       cancelText={cancelText}
       canFullscreen={canFullscreen}
+      centered={centered}
       closable={closable}
       content={children}
       draggable={draggable}
       escClosable={escClosable}
+      footer={footer}
       maskClosable={maskClosable}
       onSure={onSure}
       onCancel={onCancel}
@@ -93,6 +112,7 @@ const LModal: React.FC<SModalProps> = (props) => {
       title={title}
       transitionClass={transitionClass}
       width={width}
+      zIndex={zIndex}
     />
     , body)
 }